add inicio sesion

+++ app/Http/Controllers/AdministradorController.php
@@ -15,11 +15,19 @@ class AdministradorController extends Controller
 {
 	public function categorias(Request $request)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
     	$categorias = Categorias::paginate(10);
     	return view('categoriasAdmin', compact('categorias'));
     }
     public function crear_categoria(Request $request)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
     	$categoria = new Categorias();
         $categoria->nombre = $request->categoria;
         $categoria->activo = 0;
@@ -28,6 +36,10 @@ class AdministradorController extends Controller
     }
     public function editar_categoria(Request $request)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $categoria = Categorias::find($request->id);
         $categoria->nombre = $request->categoria;
         $categoria->save();
@@ -35,6 +47,10 @@ class AdministradorController extends Controller
     }
     public function activar_categoria($id)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $categoria = Categorias::find($id);
 		$categoria->activo = 1;
 		$categoria->save();
@@ -42,6 +58,10 @@ class AdministradorController extends Controller
     }
     public function desactivar_categoria($id)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         //desactivar productos
         $productos = Productos::where('id_categoria_fk',$id)->get();
         foreach ($productos as $producto) {
@@ -64,6 +84,10 @@ class AdministradorController extends Controller
     }
     public function eliminar_categoria($id)
     {   
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         //eliminar productos
         $productos = Productos::where('id_categoria_fk',$id)->get();
         foreach ($productos as $producto) {
@@ -83,12 +107,20 @@ class AdministradorController extends Controller
     //////////////////////////////////////////////
     public function subCategorias(Request $request)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $categorias = Categorias::All();
         $subCategorias = SubCategoria::with('categorias')->paginate(10);
         return view('subCategoriasAdmin', compact('categorias','subCategorias'));
     }
     public function crear_subCategorias(Request $request)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $subCategoria = new SubCategoria();
         $subCategoria->nombre_sub = $request->subCategoria;
         $subCategoria->id_categoria_fk = $request->id_categoria_fk;
@@ -98,6 +130,10 @@ class AdministradorController extends Controller
     }
     public function editar_subCategorias(Request $request)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $subCategoria = SubCategoria::find($request->id);
         $subCategoria->nombre_sub = $request->subCategoria;
         $subCategoria->id_categoria_fk = $request->id_categoria_fk;
@@ -106,6 +142,10 @@ class AdministradorController extends Controller
     }
     public function activar_subCategorias($id)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $subCategoria = SubCategoria::find($id);
         $subCategoria->activo_sub = 1;
         $subCategoria->save();
@@ -113,6 +153,10 @@ class AdministradorController extends Controller
     }
     public function desactivar_subCategorias($id)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         //desactivar productos
         $productos = Productos::where('id_subcategoria_fk',$id)->get();
         foreach ($productos as $producto) {
@@ -127,6 +171,10 @@ class AdministradorController extends Controller
     }
     public function eliminar_subCategorias($id)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         //eliminar productos
         $productos = Productos::where('id_categoria_fk',$id)->get();
         foreach ($productos as $producto) {
@@ -142,12 +190,20 @@ class AdministradorController extends Controller
     ////////////////////////////////////////////
     public function productos(Request $request)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $productos = Productos::with('sub_categorias')->join('categorias', 'productos.id_categoria_fk', '=', 'categorias.id')->paginate(10); // Notice the shares.* here
         $categorias = Categorias::All();
         return view('productosAdmin', compact('categorias','productos'));
     }
     public function crear_producto(Request $request)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $producto = new Productos();
         if($request->hasFile('foto')){
             $filename = 'foto'.str_random(40).".".$request->file('foto')->getClientOriginalExtension();
@@ -168,6 +224,10 @@ class AdministradorController extends Controller
     }
     public function editar_producto(Request $request)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $producto = Productos::find($request->id);
         if($request->hasFile('foto')){
             $filename = 'foto'.str_random(40).".".$request->file('foto')->getClientOriginalExtension();
@@ -186,11 +246,19 @@ class AdministradorController extends Controller
     }
     public function obtenerSubCategorias($id)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $subCategorias = SubCategoria::where('id_categoria_fk',$id)->get();
         return $subCategorias;
     }
     public function activar_producto($id)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $producto = Productos::find($id);
         $producto->activo_prod = 1;
         $producto->save();
@@ -198,6 +266,10 @@ class AdministradorController extends Controller
     }
     public function desactivar_producto($id)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $producto = Productos::find($id);
         $producto->activo_prod = 0;
         $producto->save();
@@ -205,6 +277,10 @@ class AdministradorController extends Controller
     }
     public function eliminar_producto($id)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $producto = Productos::find($id);
         File::delete('uploads/'.$producto->foto);
         $producto->delete();
@@ -212,11 +288,19 @@ class AdministradorController extends Controller
     }
     public function ventas(Request $request)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $compras = Compras::orderBy('created_at', 'desc')->paginate(10);
         return view('ventasAdmin', compact('compras'));
     }
     public function por_atender($id)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $compra = Compras::find($id);
         $compra->estado = 1;
         $compra->save();
@@ -224,14 +308,44 @@ class AdministradorController extends Controller
     }
     public function atendido($id)
     {
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
         $compra = Compras::find($id);
         $compra->estado = 0;
         $compra->save();
         return redirect()->back()->with('success', 'Compra por atender con exito');
     }
-    public function ventas(Request $request)
+    public function venta_detalle($id)
     {
-        $compras = Compras::orderBy('created_at', 'desc')->paginate(10);
-        return view('ventasAdmin', compact('compras'));
+        session_start();
+        if ((!isset($_SESSION['email_session'])) AND (!isset($_SESSION['contrasena_session']))) {
+            return redirect('/admin');
+        }
+        $compras = Compras::where('id',$id)->get();
+        return view('ventasDetalleAdmin', compact('compras'));
     }
+    public function inicioSesion()
+    {
+        return view('inicioSesion');
+    }
+    public function iniciandoSesion(Request $request)
+    {
+        if (($request->email_session=="americancompanyco") AND ($request->contrasena_session=="American0.")) {
+            session_start();
+            $_SESSION['email_session'] = "americancompanyco";
+            $_SESSION['contrasena_session'] = "American0." ;
+            return redirect('/admin/productos');
+        }else{
+            return redirect()->back();
+        }
+    }
+    public function cerrandoSesion()
+    {
+        session_start();
+        session_destroy();
+        return redirect('/admin');
+    }
+
 }
